Spawn multiple processes for multi-process consumers
After doing some digging, I found that the --single-thread AND --multi-process options are solely used for making sure that no other process is running.
So if you run a single consumer, you add the --single-thread option. With this option, Magento will check if another process has a lock for that consumer.
The same thing goes for the --multi-process n option, where Magento will check the lock in combination with specified n.