worker: Add graceful process termination and process cleanup
In the previous implementation, the processes were first signaled with
SIGTERM and almost immediately signaled with SIGKILL, not allowing the
process to stop by itself and causing potential locking problems in the
application (although not experienced as of yet).

By first signaling SIGTERM and waiting for a certain graceful period,
the process has some time to stop itself before it's forcefully killed.

This also fixes the situation where the program may have subprocesses
lingering in the Z (zombie) state, which means that a process has exited
but hasn't been waited on (with waitpid) yet.